refactor(#5): rename package name `target` to `model`-`cron/rate`
