+++ src/main/java/com/kubernetes/trafficmaker/model/TrafficTargetStatus.java
@@ -1,4 +1,4 @@
-package com.kubernetes.trafficmaker.target;
+package com.kubernetes.trafficmaker.model;
 
 public record TrafficTargetStatus(State state) {
     public enum State {
