+++ src/main/java/com/kubernetes/trafficmaker/model/rate/RateTrafficTargetSpec.java
@@ -1,0 +1,6 @@
+package com.kubernetes.trafficmaker.model.rate;
+
+import com.kubernetes.trafficmaker.model.HttpTargetSpec;
+
+public record RateTrafficTargetSpec(HttpTargetSpec http, Long rate) {
+}
